fix bugs

+++ app/src/main/java/com/mahdideebtaha/eead/utils/EnergyUtils.kt
@@ -4,11 +4,7 @@ import android.content.Context
 import android.os.BatteryManager
 
 object EnergyUtils {
-    fun averageBatteryCurrent(
-        context: Context,
-        durationMs: Long = 2000,
-        intervalMs: Long = 200
-    ): Int {
+    fun averageBatteryCurrent(context: Context, durationMs: Long = 2000, intervalMs: Long = 200): Int {
         val bm = context.getSystemService(Context.BATTERY_SERVICE) as BatteryManager
         val readings = mutableListOf<Int>()
         val start = System.currentTimeMillis()
@@ -23,6 +19,6 @@ object EnergyUtils {
     }
 
     fun calculateEnergyConsumption(avgMicroAmps: Double, durationMs: Long): Double {
-        return (avgMicroAmps / 1_000_000) * (durationMs / 1000.0 / 3600.0)
+        return (avgMicroAmps / 1_000_000) * (durationMs.toDouble() / 1000.0 / 3600.0)
     }
 }
